use `in` for ms queries + move queries to helpers
