+++ src/constants/queries/index.ts
@@ -133,102 +133,3 @@ export const getPortfolioMovements = gql`
     }
   }
 `;
-
-export const getMultisigProposalsQuery = ({
-  multisigId,
-  ids = [],
-  offset,
-  pageSize,
-  activeOnly = true,
-}: {
-  multisigId: string;
-  ids?: number[];
-  activeOnly?: boolean;
-  offset?: number;
-  pageSize?: number;
-}) => {
-  const statusFilter = activeOnly
-    ? `status: { equalTo: Active }`
-    : 'status: { notEqualTo: Active }';
-  const offsetFiler = offset ? `offset: ${offset}` : '';
-  const pageSizeFilter = pageSize ? `first: ${pageSize}` : '';
-
-  let idFilter = '';
-  if (ids.length > 0) {
-    const idFilters = ids.map((id) => `{proposalId: { equalTo: ${id} }}`);
-    idFilter = `or: [${idFilters.join(',')}]`;
-  }
-
-  const query = gql`
-    query {
-      multiSigProposals(
-        ${offsetFiler}
-        ${pageSizeFilter}
-        filter: {
-          multisigId: { equalTo: "${multisigId}" }
-          ${statusFilter}
-          ${idFilter}
-        }
-        orderBy: PROPOSAL_ID_DESC
-      ) {
-        totalCount
-        nodes {
-          approvalCount
-          createdBlockId
-          creatorAccount
-          datetime
-          extrinsicIdx
-          proposalId
-          rejectionCount
-          status
-          votes {
-            nodes {
-              action
-              signer {
-                signerValue
-              }
-            }
-          }
-        }
-      }
-    }
-  `;
-
-  return query;
-};
-
-export const getMultisigCreationExtrinsics = (
-  extrinsicArray: {
-    blockId: string;
-    extrinsicIdx: number;
-  }[],
-) => {
-  const extrinsicFilters = extrinsicArray.map(
-    ({ blockId, extrinsicIdx }) =>
-      `{ and: [
-          {blockId: { equalTo: "${blockId}" }}
-          {extrinsicIdx: { equalTo: ${extrinsicIdx} }}
-        ] }`,
-  );
-  const extrinsicFilter = `or: [${extrinsicFilters.join(',')}]`;
-
-  const query = gql`
-    query {
-      extrinsics(
-        filter: {
-          ${extrinsicFilter}
-        }
-        orderBy: CREATED_AT_DESC
-      ) {
-        totalCount
-        nodes {
-          params
-          blockId
-          extrinsicIdx
-        }
-      }
-    }
-  `;
-
-  return query;
-};
